[REFACTOR]: Refactoring code + some diagrams
