+++ diagrams/packages/controllers/auth.puml
@@ -1,0 +1,15 @@
+@startuml
+
+set separator none
+package auth.controller {
+    class AuthController <<Function Group>> {
+        + signUpWithEmail(req: Request, res: Response): RequestHandler
+        + signInWithEmail(req: Request, res: Response): RequestHandler
+        + signInWithGoogle(req: Request, res: Response): RequestHandler
+        + handleOAuthCallback(req: Request, res: Response): RequestHandler
+        + updateUserCredentials(req: Request, res: Response): RequestHandler
+        + signOut(req: Request, res: Response): RequestHandler
+    }
+}
+
+@enduml
