+++ diagrams/packages/controllers/user.puml
@@ -1,0 +1,19 @@
+@startuml
+
+set separator none
+package user.controller {
+    class UserController <<Function Group>> {
+        + getAllUsers(req: Request, res: Response): RequestHandler
+        + getUserByID(req: Request, res: Response): RequestHandler
+        + getProfile(req: Request, res: Response): RequestHandler
+        + updateProfile(req: Request, res: Response): RequestHandler
+        + getPlaylists(req: Request, res: Response): RequestHandler
+        + getListenHistory(req: Request, res: Response): RequestHandler
+        + upsertListenHistory(req: Request, res: Response): RequestHandler
+        + getFollowedArtists(req: Request, res: Response): RequestHandler
+        + followArtist(req: Request, res: Response): RequestHandler
+        + unfollowArtist(req: Request, res: Response): RequestHandler
+    }
+}
+
+@enduml
