+++ diagrams/packages/package.puml
@@ -1,0 +1,20 @@
+@startuml
+
+package controllers
+
+package middlewares
+
+package types
+
+package routes
+
+package services
+
+controllers <.. services
+controllers <.. types
+
+routes <.. controllers
+routes <.. middlewares
+
+
+@enduml
