+++ diagrams/packages/services/index.puml
@@ -1,0 +1,25 @@
+@startuml
+
+package services {
+    class Backblaze {
+        - client: S3Client {readOnly}
+        + generatePresignedDownloadURL(): Promise<string>
+        + generatePresignedUploadURL(): Promise<string>
+    }
+    class Redis {
+        - client: UpstashRedis {readOnly}
+        + set<T>(): \nPromise<"OK" | T | null>
+        + get<T>(): Promise<T | null>
+    }
+    class Cloudinary {
+        + upload(): Promise<string>
+        + delete(): Promise<void>
+    }
+    class ZaloPay {
+        - client: ZaloClient {readOnly}
+        + createOrder(): Promise<any>
+        + receiveOrderCallback(): ZaloResult
+    }
+}
+
+@enduml
