+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { JWTPayload, SanitizeOptions } from "./models/interfaces";
+import { JWTPayload, SanitizeOptions } from "@/types/interfaces";
 
 /**
  * Parse JWT payload from the Authorization header
